Evaluation of TripleObjects implemented, REPL spits out RDF/XML.
REPL can now evaluate literals, Symbols and triple statements,
manipulating a rdflib.Graph maintained in the Env class.

Next, implement a template for assignment to SYMBOLs, such that an
assignment expression is actually a triple statement adding a node for
the SYMBOL and a edge to its literal (or another URI) value.