Finished (almost) implementing QName, LocalName.
Symbol is no longer a thing. Everything is a QName, pointing to a
resource. Some work needs to be done on the REPL though - how should
the REPL display the value associated with a resource, so as to mimic
the behaviour of the old SYMBOL?

It would be ideal to keep everything just as references until the last
minute when the graph really needs to be displayed i.e. expansion
time. But this conflicts with the 'interactive' idea of the REPL.