Assignments are now parsed, implemented partially (See below)
Assignment expressions are now TopLevel expressions that evaluate as a
TripleObject. That is, an assigment like 'Variable="Value"' creates a
new rdflib.URIRef for 'Variable', and inserts a triple into the graph
with subject being the URIRef for 'Variable', predicate the special
assigment predicate URIRef returned by the evaluating environment, and
the literal "Value" as the object.

Some renaming of objects and parse rules has also been done to match
more closely with the naming conventions of ShortBOL 1.0

test_parser_literals.py is also moved to its proper home, the /test/
directory.