Grammar provided for ShortBOL 1.0
ShortBOL1/parser.py is a PLY.yacc parser generator for ShortBOL 1.0,
extracted from drdozer's shortbol/IndentingShortbolParser.scala
fastparser.

There are still some TODOs in the file, and the lexer does not provide
INDENT and DEDENT tokens, which are needed for instanceexp rules.

Note that the parser is simply a kind of reference document. The
parser builds successfully, but has not been tested, and does not
actually do anything (every almost rule has a 'pass' action associated
with it).

The plan should be to move elements of this grammar into the rdfscript
grammar slowly, checking that the results are as expected, and
implementing whatever operations possible. The end product should be a
ShortBOL 2.0 with a grammar very close to that of 1.0.