Superficial testing for parsing of the ShortBOL standard library files.
Nearly all standard library files provided in
drdozer/ShortBOLCommunity repository are parsed successfully, though
the tests are superficial, they just check that the correct 'toplevel'
forms are created by the parser.

Some files are skipped, for one of two reasons:

* They contain multi-line strings, which are not yet parsed correctly
* They contain comments in the format '// comment' which I cannot find
  in the IndentingShortbolParser from ShortBOL 1.0

Also note that QNAMEs in the stdlib files have been changed from
<prefix>:<localname> to <prefix>.<localname>, allowing <expr> to be a
toplevel form.

It is also interesting to note that these files were parsed
successfully despite the absence of the 'infixassignment' construct
defined in IndentingShortbolParser. I have not confirmed that there
are actually none in the stdlib, but if there isn't then this is good,
since 'infixassignment' causes SR conflicts with 'assignment'.