Assignment and name lookup implemented.
Identifiers LocalName and QName now have resolve() and evaluate()
methods.

resolve() returns the rdflib.URIRef for the identifier.

lookup() returns the expansion of the identifier if it exists as a
subject in the rdf graph with the environment's expansion predicate,
if not, lookup() returns None.

Testing for these changes are in test_name_evaluation.py